Update dat_num to be int matrix

+++ erpsc/count.py
@@ -68,7 +68,7 @@ class Count(Base):
         self.erp_counts = np.zeros([self.n_erps])
 
         # Initialize right size matrices to store data
-        self.dat_numbers = np.zeros([self.n_erps, self.n_terms])
+        self.dat_numbers = np.zeros([self.n_erps, self.n_terms], dtype=int)
         self.dat_percent = np.zeros([self.n_erps, self.n_terms])
 
         # Loop through each ERP term
